Padauk: Separate out parse function so it can be used with multiple chips.
