+++ configure
@@ -460,7 +460,7 @@ for option in $@; do
     NO_MSP430=""
   ;;
   --enable-padauk)
-    ASM_OBJS="${ASM_OBJS} pdk14.o"
+    ASM_OBJS="${ASM_OBJS} pdk14.o pdk_parse.o"
     DISASM_OBJS="${DISASM_OBJS} pdk14.o"
     TABLE_OBJS="${TABLE_OBJS} pdk14.o"
     DFLAGS="${DFLAGS} -DENABLE_PADAUK"
